requirements.txt
fastapi
uvicorn[standard]
requests
numpy
supabase
python-telegram-bot
python-dotenv
